Add starfield, fix warnings

+++ src/Globe.cpp
@@ -16,7 +16,7 @@ void Globe::update() {
 	// Copy spectrum to a texture for the shader to use
 	if (analysis) {
 		ranges = analysis->getHPCP();
-		spectrumTex.loadData(ranges.data(), ranges.size(), 1, GL_RED); // GL_RED is single value in red channel
+		spectrumTex.loadData(ranges.data(), static_cast<int>(ranges.size()), 1, GL_RED); // GL_RED is single value in red channel
 
 		if (!didSetVertFrequencies) {
 			calcVertexFrequencies();
@@ -31,7 +31,7 @@ void Globe::customDraw() {
 	auto& shader = material.getShader();
 	shader.setUniform1f("amplitude", amplitude / radius); // amplitude is relative to scale
 	shader.setUniform1f("colorHueRange", colorHueRange);
-	shader.setUniform1i("bins", ranges.size());
+	shader.setUniform1i("bins", static_cast<int>(ranges.size()));
 	if (spectrumTex.isAllocated()) shader.setUniformTexture("spectrum", spectrumTex, 1); // 1-indexed
 
 	ofPushMatrix();
@@ -55,7 +55,7 @@ void Globe::calcVertexFrequencies() {
 	const float NOISE_SCALE = 200;
 
 	std::vector<float> vertFrequency(mesh.getNumVertices());
-	for (size_t i = 0; i < mesh.getNumVertices(); i++) {
+	for (int i = 0; i < mesh.getNumVertices(); i++) {
 		ofVec3f pos = mesh.getVertex(i);
 		vertFrequency[i] = ofNoise(pos * NOISE_SCALE) * ranges.size(); // needs to be between 0 and spectrum texture width
 	}
@@ -63,5 +63,5 @@ void Globe::calcVertexFrequencies() {
 	// Link the frequencies to the mesh
 	auto& shader = material.getShader();
 	GLint freqAttrib = shader.getAttributeLocation("frequency");
-	mesh.getVbo().setAttributeData(freqAttrib, vertFrequency.data(), 1, mesh.getNumVertices(), GL_STATIC_DRAW, sizeof(float));
+	mesh.getVbo().setAttributeData(freqAttrib, vertFrequency.data(), 1, static_cast<int>(mesh.getNumVertices()), GL_STATIC_DRAW, sizeof(float));
 }
